+++ lizard_auth_client/views.py
@@ -19,8 +19,14 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth.decorators import permission_required
 from django.contrib.auth.models import User
 from django.core.exceptions import ObjectDoesNotExist
-from django.core.urlresolvers import reverse
-from django.core.urlresolvers import reverse_lazy
+try:
+    from django.core.urlresolvers import reverse
+    from django.core.urlresolvers import reverse_lazy
+# django > 1.9 combat
+except ImportError:
+    from django.urls import reverse
+    from django.urls import reverse_lazy
+
 from django.db import transaction
 from django.forms.models import model_to_dict
 from django.http import Http404